Fix process CPU usage becomes 0 bug
The dynamic changing in small CPU usage process is also very small.
Therefore, it shouldn't be updating so frequently in small CPU usage
process.
There is also a bug in the initialization of previous uptime in process.
It should be initialize when the process is created, not when it is 0.